Core/Misc: Future msvc warning fixes

+++ src/server/game/Spells/SpellInfo.cpp
@@ -389,7 +389,7 @@ bool SpellEffectInfo::IsEffect() const
 
 bool SpellEffectInfo::IsEffect(SpellEffects effectName) const
 {
-    return Effect == uint32(effectName);
+    return Effect == effectName;
 }
 
 bool SpellEffectInfo::IsAura() const
